2012-01-31 10:00 UTC+0700 Andi Jahja <xharbour/AT/telkom/net/id>   * source/rtl/zlib/adler32.c   * source/rtl/zlib/compress.c   * source/rtl/zlib/crc32.c   * source/rtl/zlib/crc32.h   * source/rtl/zlib/deflate.c   * source/rtl/zlib/deflate.h   * source/rtl/zlib/infback.c   * source/rtl/zlib/inffast.c   * source/rtl/zlib/inffast.h   * source/rtl/zlib/inffixed.h   * source/rtl/zlib/inflate.c   * source/rtl/zlib/inflate.h   * source/rtl/zlib/inftrees.c   * source/rtl/zlib/inftrees.h   * source/rtl/zlib/trees.c   * source/rtl/zlib/trees.h   * source/rtl/zlib/uncompr.c   * source/rtl/zlib/zconf.h   * source/rtl/zlib/zconf.in.h   * source/rtl/zlib/zlib.h   * source/rtl/zlib/zutil.c   * source/rtl/zlib/zutil.h   - source/rtl/zlib/gzio.c   + source/rtl/zlib/gzlib.c   + source/rtl/zlib/gzwrite.c   + source/rtl/zlib/gzguts.h   * winmake/common.mak   * winmake/compile.mak     ! updated to zlib v. 1.26
  * contrib/png/png.c
  * contrib/png/png.h
  * contrib/png/pngconf.h
  * contrib/png/pngdebug.h
  * contrib/png/pngerror.c
  * contrib/png/pngget.c
  * contrib/png/pnginfo.h
  * contrib/png/pnglibconf.h
  * contrib/png/pngmem.c
  * contrib/png/pngpread.c
  * contrib/png/pngpriv.h
  * contrib/png/pngread.c
  * contrib/png/pngrio.c
  * contrib/png/pngrtran.c
  * contrib/png/pngrutil.c
  * contrib/png/pngset.c
  * contrib/png/pngstruct.h
  * contrib/png/pngtrans.c
  * contrib/png/pngwio.c
  * contrib/png/pngwrite.c
  * contrib/png/pngwtran.c
  * contrib/png/pngwutil.c
    ! updated to png v. 1.57

  * source/rtl/pcre/config.h
  * source/rtl/pcre/config.h.generic
  * source/rtl/pcre/pcre.h
  * source/rtl/pcre/pcre.h.generic
  * source/rtl/pcre/pcreposix.c
  * source/rtl/pcre/pcreposix.h
  * source/rtl/pcre/pcre_chartables.c
  * source/rtl/pcre/pcre_chartables.c.dist
  * source/rtl/pcre/pcre_compile.c
  * source/rtl/pcre/pcre_config.c
  * source/rtl/pcre/pcre_dfa_exec.c
  * source/rtl/pcre/pcre_exec.c
  * source/rtl/pcre/pcre_fullinfo.c
  * source/rtl/pcre/pcre_get.c
  * source/rtl/pcre/pcre_globals.c
  * source/rtl/pcre/pcre_info.c
  * source/rtl/pcre/pcre_internal.h
  * source/rtl/pcre/pcre_maketables.c
  * source/rtl/pcre/pcre_newline.c
  * source/rtl/pcre/pcre_ord2utf8.c
  * source/rtl/pcre/pcre_printint.src
  * source/rtl/pcre/pcre_refcount.c
  * source/rtl/pcre/pcre_scanner.h
  * source/rtl/pcre/pcre_study.c
  * source/rtl/pcre/pcre_tables.c
  * source/rtl/pcre/pcre_try_flipped.c
  * source/rtl/pcre/pcre_ucd.c
  * source/rtl/pcre/pcre_valid_utf8.c
  * source/rtl/pcre/pcre_version.c
  * source/rtl/pcre/pcre_xclass.c
  * source/rtl/pcre/ucp.h
    ! updated to pcre v. 8.21

  * utils/hbdoc/hbdoc.prg
  * utils/hbpp/hbpp.c
  * utils/hbmake/hbmake.prg
  * utils/hbtest/hbtest.prg
  * utils/hbrun/hbrun.prg
  * compiler/hbusage.c
    ! bump copyright years to 2012

+++ xharbour/source/rtl/pcre/pcre_dfa_exec.c
@@ -2784,7 +2784,7 @@ for (;;)
             {
             const uschar *p = ptr;
             const uschar *pp = local_ptr;
-            charcount = pp - p;
+            charcount = (int)(pp - p);
             while (p < pp) if ((*p++ & 0xc0) == 0x80) charcount--;
             ADD_NEW_DATA(-next_state_offset, 0, (charcount - 1));
             }
